requirements.txt
lxml==4.9.3
Pillow==10.0.0
python-pptx==0.6.18
XlsxWriter==3.1.2
